+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/tobe/ControllerHandlerAdapterTest.java
@@ -7,21 +7,22 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.mvc.asis.ControllerHandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
 import webmvc.org.springframework.web.servlet.view.JspView;
 
 @SuppressWarnings("NonAsciiCharacters")
-@DisplayName("ManualHandlerAdapter 테스트")
+@DisplayName("ControllerHandlerAdapter 테스트")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class ManualHandlerAdapterTest {
+class ControllerHandlerAdapterTest {
 
     @Test
     void 컨트롤러_인터페이스를_구현한_핸들러만_지원한다() {
         // given
-        final ManualHandlerAdapter manualHandlerAdapter = new ManualHandlerAdapter();
+        final ControllerHandlerAdapter controllerHandlerAdapter = new ControllerHandlerAdapter();
 
         // when
-        final boolean isSupport = manualHandlerAdapter.isSupport(new ForwardController("/"));
+        final boolean isSupport = controllerHandlerAdapter.isSupport(new ForwardController("/"));
 
         // then
         assertThat(isSupport).isTrue();
@@ -30,10 +31,10 @@ class ManualHandlerAdapterTest {
     @Test
     void 컨트롤러_인터페이스를_구현하지_않은_핸들러는_지원하지_않는다() {
         // given
-        final ManualHandlerAdapter manualHandlerAdapter = new ManualHandlerAdapter();
+        final ControllerHandlerAdapter controllerHandlerAdapter = new ControllerHandlerAdapter();
 
         // when
-        final boolean isSupport = manualHandlerAdapter.isSupport(new Object());
+        final boolean isSupport = controllerHandlerAdapter.isSupport(new Object());
 
         // then
         assertThat(isSupport).isFalse();
@@ -42,11 +43,11 @@ class ManualHandlerAdapterTest {
     @Test
     void 컨트롤러_인터페이스를_구현한_핸들러를_실행한다() throws Exception {
         // given
-        final ManualHandlerAdapter manualHandlerAdapter = new ManualHandlerAdapter();
+        final ControllerHandlerAdapter controllerHandlerAdapter = new ControllerHandlerAdapter();
         final ForwardController forwardController = new ForwardController("/index.jsp");
 
         // when
-        final ModelAndView modelAndView = manualHandlerAdapter.handle(null, null, forwardController);
+        final ModelAndView modelAndView = controllerHandlerAdapter.handle(null, null, forwardController);
 
         // then
         assertThat(modelAndView.getView()).isInstanceOf(JspView.class);
